+++ mongod/update.go
@@ -202,7 +202,7 @@ func (c *appRepository) UpdateArch(id primitive.ObjectID, archID string, owner s
 }
 
 // UpdateApp updates an existing app_name document
-func (c *appRepository) UpdateApp(id primitive.ObjectID, appName string, logo string, tuf bool, description string, reports bool, cdnEdge bool, owner string, ctx context.Context) (interface{}, error) {
+func (c *appRepository) UpdateApp(id primitive.ObjectID, appName string, logo string, tuf bool, description string, reports bool, cdnEdge bool, downloadMode string, owner string, ctx context.Context) (interface{}, error) {
 	filter := bson.D{{Key: "_id", Value: id}}
 	updateFields := bson.D{{Key: "app_name", Value: appName}}
 	if logo != "" {
@@ -214,6 +214,9 @@ func (c *appRepository) UpdateApp(id primitive.ObjectID, appName string, logo st
 	updateFields = append(updateFields, bson.E{Key: "tuf", Value: tuf})
 	updateFields = append(updateFields, bson.E{Key: "reports", Value: reports})
 	updateFields = append(updateFields, bson.E{Key: "cdn_edge", Value: cdnEdge})
+	if downloadMode != "" {
+		updateFields = append(updateFields, bson.E{Key: "download_mode", Value: downloadMode})
+	}
 	update := bson.D{{Key: "$set", Value: updateFields}}
 	return c.UpdateDocument("apps_meta", filter, update, "app_name_sort_by_asc_updated", "app", owner, ctx)
 }
@@ -426,6 +429,7 @@ func (c *appRepository) UpdateSpecificApp(objID primitive.ObjectID, owner string
 			if !duplicateFound {
 				newArtifact := model.Artifact{
 					Link:      appLink,
+					S3Key:     utils.PrivateObjectKey(appLink),
 					Platform:  platformMeta.ID,
 					Arch:      archMeta.ID,
 					Package:   extension,